Saving Examinations

+++ src/main/java/pl/softsystem/hospital/repository/ExaminationRepository.java
@@ -3,5 +3,5 @@ package pl.softsystem.hospital.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.softsystem.hospital.model.Examination;
 
-public interface ExaminationRepository extends JpaRepository<Examination, Integer> {
+public interface ExaminationRepository extends JpaRepository<Examination, Long> {
 }
