+++ src/main/java/pl/softsystem/hospital/repository/ResultRepository.java
@@ -3,5 +3,5 @@ package pl.softsystem.hospital.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.softsystem.hospital.model.Result;
 
-public interface ResultRepository extends JpaRepository<Result, Integer> {
+public interface ResultRepository extends JpaRepository<Result, Long> {
 }
